Compiler warnings

+++ mops/src/hpe/cpu.tpp
@@ -34,18 +34,18 @@ void mops::homogeneous_polynomial_evaluation(
     scalar_t* c_ptr = tensor_c.data;
     int32_t* p_ptr = p.data;
 
-    long size_batch_dimension = tensor_a.shape[0];
-    long n_monomials = p.shape[0];
-    long polynomial_order = p.shape[1];
-    long n_possible_factors = tensor_a.shape[1];
+    size_t size_batch_dimension = tensor_a.shape[0];
+    size_t n_monomials = p.shape[0];
+    size_t polynomial_order = p.shape[1];
+    size_t n_possible_factors = tensor_a.shape[1];
 
-    for (long i = 0; i < size_batch_dimension; i++) {
+    for (size_t i = 0; i < size_batch_dimension; i++) {
         scalar_t result = 0.0;
         scalar_t* shifted_a_ptr = a_ptr + i*n_possible_factors;
         int32_t* p_ptr_temp = p_ptr;
-        for (long j = 0; j < n_monomials; j++) {
+        for (size_t j = 0; j < n_monomials; j++) {
             scalar_t temp = c_ptr[j];
-            for (uint8_t k = 0; k < polynomial_order; k++) {
+            for (size_t k = 0; k < polynomial_order; k++) {
                 temp *= shifted_a_ptr[p_ptr_temp[k]];
             }
             result += temp;
